Add option to make the leading stars invisible
From  https://github.com/sabof/org-bullets/pull/13. I wanted to remove
the indentation from the org levels and this pull request against the
original repository did this. I however had to make the changes to this
file by hand because the emacsorphanage version has added most of the
stuff that used to be in the minor mode definition in the
org-bullets--keywords variable instead and there would thus be a
conflict if I attempted to merge the specific commits.